Debug

+++ scripts/rtp_install.bash
@@ -12,26 +12,28 @@ if [ ! -d RTP* ] ; then
   unar -e cp932 2000rtp.zip
 fi
 
-export LANG=ja_JP.UTF-8
-
-RTP_EXE=$(find RTP* -name RPG2000RTP.exe)
-cp ${RTP_EXE} ./RPG2000RTP.exe
+if [ ! -f RTP2000RTP.exe ] ; then
+  RTP_EXE=$(find RTP* -name RPG2000RTP.exe)
+  cp ${RTP_EXE} ./RPG2000RTP.exe
+fi
 
 if [ ! -v WINEPREFIX ] ; then
   export WINEPREFIX=$HOME/.wine
 fi
 
 export DISPLAY=:1024
-Xvfb $DISPLAY -screen 0 1920x1080x24 &
+Xvfb "${DISPLAY}" -screen 0 1920x1080x24 &
 
-sleep 0.1
+export LANG=ja_JP.UTF-8
 
 winecfg /v win10
-cp setup.iss "${WINEPREFIX}/drive_c/setup.iss"
+cp setup.iss "${WINEPREFIX}/drive_c"
+
+export WINEDEBUG=warn+all
+wine cmd.exe /c echo test
 
 wine ./RPG2000RTP.exe /s /a /s /sms /f1C:\\setup.iss
 
-find "$WINEPREFIX/drive_c"
-ls "$WINEPREFIX/drive_c/Program Files (x86)/ASCII"
+ls "${WINEPREFIX}/drive_c/Program Files (x86)/ASCII"
 
 kill %1
